MIDI playback support, refactoring - added support for MIDI playback with timidity - general code cleanup/refactoring
