+++ BUGS.md
@@ -1,0 +1,3 @@
+- mediainfo reports incorrect sampel counts/duration. disabled for the time being.
+- ffmpeg seeks modfiles incorrectly. considering switching to xmp for rendering for the time being.
+	- looks like timidity supports mod playback too
