+++ README.md
@@ -1,2 +1,30 @@
-# audio-player
-audio playback module for python
+### Introduction
+TODO
+
+### Installation
+`pip install git+http://github.com/0xf0f/audio-player`
+https://github.com/adaptlearning/adapt_authoring/wiki/Installing-FFmpeg
+
+timidity
+
+sox
+
+### Examples
+TODO
+
+### Documentation
+TODO
+
+### Credits
+- soundfile
+- sounddevice
+- samplerate
+- numba
+- numpy
+- cached_property
+
+- timidity
+- ffmpeg
+- loop.wav/loop2.wav
+- sox
+- pyqt5
