+++ audio_player/__init__.py
@@ -1,3 +1,4 @@
 from .audio_player import AudioPlayer
 from .audio_player_process_interface import AudioPlayerProcessInterface
+from .lib.exceptions import UnableToOpenFileError
 # from .audio_player_process import AudioPlayerProcess
