+++ audio_player/audio_player_process.py
@@ -51,13 +51,12 @@ class AudioPlayerProcess(mp.Process):
         self.audio_player.resume()
 
     def run(self) -> None:
-        from audio_player.lib.code_timer import CodeTimer
-
-        with CodeTimer('process init'):
-            self.process_init()
+        self.process_init()
 
         while True:
             command, args = self.command_queue.get()
-            # print(command, args)
-            command = self.command_map.get(command, lambda *args, **kwargs: None)
+            command = self.command_map.get(
+                command, lambda *args, **kwargs: None
+            )
+
             command(*args)
