+++ audio_player/lib/audio_file/adapters/ffmpeg/__init__.py
@@ -1,0 +1,1 @@
+from .ffmpeg_adapter import FFMPEGAdapter
