+++ audio_player/lib/audio_file/adapters/soundfile/__init__.py
@@ -1,0 +1,1 @@
+from .soundfile_adapter import SoundFileAdapter
