+++ audio_player/lib/audio_file/adapters/timidity/__init__.py
@@ -1,0 +1,1 @@
+from .timidity_adapter import TiMidityAdapter
