+++ audio_player/lib/audio_file/adapters/timidity/default.cfg
@@ -1,0 +1,1 @@
+soundfont "sYnerGi-8Mb.sf2"
