+++ audio_player/lib/audio_info/adapters/ffprobe/__init__.py
@@ -1,0 +1,1 @@
+from .ffprobe_adapter import FFProbeAdapter
