--- audio_player/lib/audio_info/adapters/ffprobe_process.py
@@ -1,17 +1,0 @@
-import subprocess as sp
-import json
-
-
-def probe(path):
-    process = sp.Popen(
-        [
-            'ffprobe', path,
-            '-print_format', 'json=c=1',
-            '-show_streams', '-show_format',
-            '-loglevel', 'quiet'
-        ],
-
-        stdout=sp.PIPE, stderr=sp.DEVNULL,
-    )
-
-    return json.load(process.stdout)
