+++ audio_player/lib/audio_info/adapters/mediainfo/__init__.py
@@ -1,0 +1,1 @@
+from .mediainfo_adapter import MediaInfoAdapter
