+++ audio_player/lib/exceptions/__init__.py
@@ -1,0 +1,3 @@
+class UnableToOpenFileError(Exception):
+    def __init__(self, path):
+        super().__init__(f'Error occurred while opening {path}')
